RDM-7627 actioned review comments
